refactor(tsdb): move series file config to seriesfile package
